Remove styled-components and react-router-dom to simplify frontend.
