+++ frontend/src/features/classifier/styles.css
@@ -1,0 +1,27 @@
+video {
+  border: 1px solid black;
+  margin: 0;
+  max-width: 90%;
+}
+
+/* Hide canvas used for capturing images */
+canvas {
+  display: none;
+}
+
+.camera-wrapper {
+  align-content: center;
+  height: auto;
+  margin: 0;
+  padding: 0;
+}
+
+.compatNotice {
+  font-size: 12px;
+  font-style: italic;
+}
+
+.result-text {
+  font-size: 1.5rem;
+  font-weight: bold;
+}
