--- frontend/src/pages/Home/styled.jsx
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const LoadingTextWrapper = styled.p`
-  font-style: italic;
-`;
-
-export const TimeWrapper = styled.p`
-  font-weight: bold;
-`;
