--- frontend/src/styled.jsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const GlobalStyles = styled.div`
-  font-family: 'Verdana', sans-serif;
-  text-align: center;
-  margin: 0;
-  padding: 0;
-`;
