+++ frontend/src/styles.css
@@ -1,3 +1,6 @@
 body {
-    margin: 0;
+  font-family: 'Verdana', sans-serif;
+  margin: 0;
+  padding: 0;
+  text-align: center;
 }
