+++ package.json
@@ -10,9 +10,7 @@
   "dependencies": {
     "@babel/runtime": "^7.11.2",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "styled-components": "^5.2.0"
+    "react-dom": "^16.13.1"
   },
   "devDependencies": {
     "@babel/core": "^7.11.6",
